+++ Meta-Doc-main/APIMetaDoc/Controllers/ProductController.cs
@@ -37,12 +37,18 @@ namespace APIMetaDoc.Controllers
         [Logged]
         [HttpGet]
         [Route("api/products/{id}")]
-        public HttpResponseMessage Products(int Id)
+        public HttpResponseMessage Products(int Id) //pharmacy id
         {
             try
             {
-                var data = ProductService.Get(Id);
-                return Request.CreateResponse(HttpStatusCode.OK, data);
+                var data1 = PharProductService.Get();
+                var status = data1.FirstOrDefault(x => x.Pharmacy_Id == Id);
+                if (status != null)
+                {
+                    var data = ProductService.Get(status.Product_Id);
+                    return Request.CreateResponse(HttpStatusCode.OK, data);
+                }
+                return Request.CreateResponse(HttpStatusCode.OK, new { Message = "Product Not Accessible" });
             }
             catch (Exception ex)
             {
@@ -80,9 +86,17 @@ namespace APIMetaDoc.Controllers
             {
                 try
                 {
-                    var res = ProductService.Update(data);
-
-                    return Request.CreateResponse(HttpStatusCode.OK, new { Message = "Updated" });
+                    var data1 = PharProductService.Get();
+                    var status = data1.FirstOrDefault(x => x.Product_Id == exmp.Id);
+                    if (status != null)
+                    {
+                        var res = ProductService.Update(exmp);
+                        return Request.CreateResponse(HttpStatusCode.OK, new { Message = "Updated" });
+                    }
+                    else
+                    {
+                        return Request.CreateResponse(HttpStatusCode.BadRequest, new { Message = "You Can't Update Others Product" });
+                    }
 
                 }
                 catch (Exception ex)
@@ -106,8 +120,16 @@ namespace APIMetaDoc.Controllers
             {
                 try
                 {
-                    var res = ProductService.Delete(Id);
-                    return Request.CreateResponse(HttpStatusCode.OK, new { Message = "Deleted" });
+                    var data = PharProductService.Get();
+                    var status = data.FirstOrDefault(x => x.Product_Id == exmp.Id);
+                    if (status != null)
+                    {
+                        var res = ProductService.Delete(Id);
+                        return Request.CreateResponse(HttpStatusCode.OK, new { Message = "Deleted" });
+                    } else
+                    {
+                        return Request.CreateResponse(HttpStatusCode.BadRequest, new { Message = "You Can't Delete Others Product" });
+                    }
 
                 }
                 catch (Exception ex)
